Create supplement
Studies that relate colonization of sites in a metapopulation to measures of site connectivity (or it’s inverse, isolation) often ignore two forms of uncertainty–the occupancy status of sites that have not been surveyed, and the chance that sites where the species of interest was not detected could have been occupied by the species. The approach we develop makes use of two ways that site occupancy (or the probability that a species was present at a site but never detected in a season) can be calculated–unconditional and conditional occupancy (MacKenzie et al. 2006). It builds on the advances of Royle and Dorazio (2008), Yackulic et al. (2012) and Eaton et al. (2014) that used an autoregressive function in occupancy models to address similar issues.   

The model of metapopulation dynamics uses MacKenzie et al.’s (2003) single-species multi-season occupancy formulation, which uses binary detection histories with revisits within a season and assumes no false positives (i.e., 11 00 01 for a patch where the species was detected on both visits in season one, not detected in season two, and detected on the second visit only in season three) to estimate detection-corrected patch unconditional occupancy in the first season (ψ[i,1]) and patch colonization (γ[i,t]) and extinction (ε[i,t]) in subsequent seasons. Repeated visits within a season are used to estimate the probability of detection per visit j at patch i in season t (p[i,t,j]). Unconditional occupancy in seasons after the first can be calculated as a derived parameter as the probability that the site was previous occupied and did not go extinct, or was previously unoccupied and was colonized:

ψ[i,t] = ψ[i,t-1] * (1-ε[i,t]) + (1 - ψ[i,t-1]) * γ[i,t]